# colinfo: optimize and refactor ResultColumns.NodeFormatter