--- a/pkg/sql/filter.go
+++ b/pkg/sql/filter.go
@@ -45,7 +45,7 @@ func (f *filterNode) IndexedVarResolvedType(idx int) *types.T {
 
 // IndexedVarNodeFormatter implements the tree.IndexedVarContainer interface.
 func (f *filterNode) IndexedVarNodeFormatter(idx int) tree.NodeFormatter {
-	return f.source.columns.NodeFormatter(idx)
+	return f.source.columns.Name(idx)
 }
 
 func (f *filterNode) startExec(runParams) error {