#### colinfo: eliminate allocations in ResultColumns.NodeFormatter

This commit eliminates allocations in `ResultColumns.NodeFormatter` by
returning an already-allocated `*string` casted to a `*tree.Name`.

Release notes: None

#### colinfo: refactor and rename ResultColumns.NodeFormatter

This commit replaces `ResultColumns.NodeFormatter` with
`ResultColumns.Name` with the return type of `*tree.Name` instead of
`tree.NodeFormatter`. `scanNode.IndexedVarNodeFormatter` has been
updated to use this new `Name` method.

Epic: None

Release note: None