@mgartner 2024-01-06
I've seen these allocations responsible for 5-10% of all allocations in some real-world profiles. They scale with the number of columns flowing through a query plan. The end-to-end benchmarks with many columns show the benefit:

```
name                                              old time/op    new time/op    delta
EndToEnd/kv-read/Simple-10                           145µs ± 1%     150µs ± 5%    ~     (p=0.690 n=5+5)
EndToEnd/kv-read/Prepared-10                        88.6µs ± 3%    87.4µs ± 2%    ~     (p=0.310 n=5+5)
EndToEnd/kv-read-const/Simple-10                     100µs ± 1%      99µs ± 2%    ~     (p=0.421 n=5+5)
EndToEnd/kv-read-const/Prepared-10                   112µs ±33%     114µs ±14%    ~     (p=0.548 n=5+5)
EndToEnd/tpcc-new-order/Simple-10                    185µs ± 1%     198µs ±21%    ~     (p=0.690 n=5+5)
EndToEnd/tpcc-new-order/Prepared-10                  100µs ± 3%     102µs ± 3%    ~     (p=0.548 n=5+5)
EndToEnd/tpcc-delivery/Simple-10                     245µs ± 1%     255µs ±10%    ~     (p=0.690 n=5+5)
EndToEnd/tpcc-delivery/Prepared-10                   156µs ± 4%     152µs ± 2%  -2.80%  (p=0.032 n=5+5)
EndToEnd/tpcc-stock-level/Simple-10                  493µs ± 4%     527µs ±17%    ~     (p=0.310 n=5+5)
EndToEnd/tpcc-stock-level/Prepared-10                304µs ± 1%     303µs ± 1%    ~     (p=0.222 n=5+5)
EndToEnd/many-columns-and-indexes-a/Simple-10        680µs ± 0%     685µs ± 3%    ~     (p=0.548 n=5+5)
EndToEnd/many-columns-and-indexes-a/Prepared-10      579µs ± 1%     581µs ± 1%    ~     (p=0.548 n=5+5)
EndToEnd/many-columns-and-indexes-b/Simple-10        713µs ± 1%     735µs ± 1%  +3.04%  (p=0.008 n=5+5)
EndToEnd/many-columns-and-indexes-b/Prepared-10      594µs ± 1%     588µs ± 3%    ~     (p=0.310 n=5+5)
EndToEnd/many-columns-and-indexes-c/Simple-10        178ms ± 1%     179ms ± 1%    ~     (p=0.686 n=4+4)
EndToEnd/many-columns-and-indexes-c/Prepared-10      177ms ± 2%     178ms ± 0%    ~     (p=0.151 n=5+5)
EndToEnd/single-col-histogram-range/Simple-10        229µs ± 1%     228µs ± 1%    ~     (p=0.095 n=5+5)
EndToEnd/single-col-histogram-range/Prepared-10      137µs ± 3%     130µs ± 2%  -5.80%  (p=0.008 n=5+5)
EndToEnd/batch-insert-one/Simple-10                 29.0ms ± 4%    28.6ms ± 1%    ~     (p=0.222 n=5+5)
EndToEnd/batch-insert-one/Prepared-10               28.8ms ± 1%    28.8ms ± 1%    ~     (p=0.690 n=5+5)
EndToEnd/batch-insert-many/Simple-10                34.5ms ± 2%    34.2ms ± 2%    ~     (p=0.690 n=5+5)
EndToEnd/batch-insert-many/Prepared-10              35.6ms ±15%    35.4ms ±14%    ~     (p=0.421 n=5+5)
EndToEnd/ored-preds-100/Simple-10                   1.44ms ± 3%    1.42ms ± 0%  -1.78%  (p=0.032 n=5+5)
EndToEnd/ored-preds-100/Prepared-10                  529µs ± 3%     513µs ± 6%    ~     (p=0.151 n=5+5)
EndToEnd/ored-preds-using-params-100/Simple-10      2.52ms ± 1%    2.48ms ± 1%  -1.83%  (p=0.016 n=5+5)
EndToEnd/ored-preds-using-params-100/Prepared-10    1.37ms ± 2%    1.32ms ± 1%  -3.34%  (p=0.008 n=5+5)

name                                              old alloc/op   new alloc/op   delta
EndToEnd/kv-read/Simple-10                          39.2kB ± 4%    39.0kB ± 4%    ~     (p=0.841 n=5+5)
EndToEnd/kv-read/Prepared-10                        26.8kB ± 0%    27.2kB ± 3%    ~     (p=1.000 n=4+5)
EndToEnd/kv-read-const/Simple-10                    27.8kB ± 4%    27.7kB ± 4%    ~     (p=1.000 n=5+5)
EndToEnd/kv-read-const/Prepared-10                  26.9kB ± 3%    27.1kB ± 3%    ~     (p=0.690 n=5+5)
EndToEnd/tpcc-new-order/Simple-10                   45.1kB ± 7%    44.9kB ± 8%    ~     (p=0.421 n=5+5)
EndToEnd/tpcc-new-order/Prepared-10                 28.9kB ± 3%    28.8kB ± 3%    ~     (p=0.690 n=5+5)
EndToEnd/tpcc-delivery/Simple-10                    63.3kB ± 4%    63.1kB ± 4%    ~     (p=0.548 n=5+5)
EndToEnd/tpcc-delivery/Prepared-10                  47.3kB ± 0%    47.3kB ± 0%    ~     (p=0.413 n=4+5)
EndToEnd/tpcc-stock-level/Simple-10                  204kB ±64%     219kB ±65%    ~     (p=0.841 n=5+5)
EndToEnd/tpcc-stock-level/Prepared-10                143kB ± 3%     141kB ± 0%    ~     (p=0.063 n=5+4)
EndToEnd/many-columns-and-indexes-a/Simple-10       60.1kB ± 0%    59.5kB ± 0%  -1.05%  (p=0.008 n=5+5)
EndToEnd/many-columns-and-indexes-a/Prepared-10     49.8kB ±16%    48.1kB ± 2%    ~     (p=0.730 n=5+4)
EndToEnd/many-columns-and-indexes-b/Simple-10       64.5kB ± 1%    64.2kB ± 1%    ~     (p=0.486 n=4+4)
EndToEnd/many-columns-and-indexes-b/Prepared-10     51.4kB ± 1%    50.9kB ± 0%  -1.07%  (p=0.008 n=5+5)
EndToEnd/many-columns-and-indexes-c/Simple-10        107MB ± 2%     112MB ±17%    ~     (p=0.905 n=4+5)
EndToEnd/many-columns-and-indexes-c/Prepared-10      106MB ± 0%     106MB ± 0%    ~     (p=0.886 n=4+4)
EndToEnd/single-col-histogram-range/Simple-10       43.1kB ± 0%    43.0kB ± 0%    ~     (p=0.222 n=5+5)
EndToEnd/single-col-histogram-range/Prepared-10     33.6kB ±16%    32.9kB ± 6%    ~     (p=1.000 n=5+5)
EndToEnd/batch-insert-one/Simple-10                 6.89MB ±11%    6.86MB ±10%    ~     (p=1.000 n=5+5)
EndToEnd/batch-insert-one/Prepared-10               6.85MB ± 7%    7.17MB ±14%    ~     (p=0.222 n=5+5)
EndToEnd/batch-insert-many/Simple-10                14.6MB ±20%    13.9MB ± 1%    ~     (p=1.000 n=5+4)
EndToEnd/batch-insert-many/Prepared-10              11.2MB ± 4%    13.3MB ±52%    ~     (p=0.556 n=4+5)
EndToEnd/ored-preds-100/Simple-10                    446kB ± 0%     430kB ± 0%  -3.61%  (p=0.029 n=4+4)
EndToEnd/ored-preds-100/Prepared-10                  277kB ± 0%     261kB ± 0%  -5.87%  (p=0.016 n=4+5)
EndToEnd/ored-preds-using-params-100/Simple-10      1.20MB ± 0%    1.19MB ± 0%  -1.39%  (p=0.008 n=5+5)
EndToEnd/ored-preds-using-params-100/Prepared-10    1.03MB ± 0%    1.02MB ± 0%  -1.54%  (p=0.016 n=4+5)

name                                              old allocs/op  new allocs/op  delta
EndToEnd/kv-read/Simple-10                             424 ± 3%       421 ± 3%    ~     (p=0.563 n=5+5)
EndToEnd/kv-read/Prepared-10                           293 ± 0%       293 ± 0%    ~     (p=1.000 n=5+5)
EndToEnd/kv-read-const/Simple-10                       306 ± 3%       302 ± 0%    ~     (p=0.159 n=5+4)
EndToEnd/kv-read-const/Prepared-10                     286 ± 0%       289 ± 2%    ~     (p=0.556 n=4+5)
EndToEnd/tpcc-new-order/Simple-10                      458 ± 0%       456 ± 0%    ~     (p=0.057 n=4+4)
EndToEnd/tpcc-new-order/Prepared-10                    320 ± 1%       317 ± 1%  -0.94%  (p=0.024 n=5+5)
EndToEnd/tpcc-delivery/Simple-10                       630 ± 0%       630 ± 0%    ~     (p=0.171 n=4+4)
EndToEnd/tpcc-delivery/Prepared-10                     466 ± 0%       465 ± 0%    ~     (p=0.429 n=4+4)
EndToEnd/tpcc-stock-level/Simple-10                  1.50k ±52%     1.59k ±53%    ~     (p=0.889 n=5+5)
EndToEnd/tpcc-stock-level/Prepared-10                1.05k ± 0%     1.05k ± 0%  -0.26%  (p=0.029 n=4+4)
EndToEnd/many-columns-and-indexes-a/Simple-10          544 ± 0%       540 ± 0%  -0.70%  (p=0.008 n=5+5)
EndToEnd/many-columns-and-indexes-a/Prepared-10        424 ± 0%       422 ± 1%    ~     (p=0.457 n=4+4)
EndToEnd/many-columns-and-indexes-b/Simple-10          592 ± 0%       588 ± 0%  -0.68%  (p=0.029 n=4+4)
EndToEnd/many-columns-and-indexes-b/Prepared-10        466 ± 0%       461 ± 0%  -1.03%  (p=0.008 n=5+5)
EndToEnd/many-columns-and-indexes-c/Simple-10         718k ± 2%      718k ± 2%    ~     (p=0.886 n=4+4)
EndToEnd/many-columns-and-indexes-c/Prepared-10       712k ± 0%      712k ± 0%    ~     (p=0.486 n=4+4)
EndToEnd/single-col-histogram-range/Simple-10          436 ± 0%       436 ± 0%    ~     (p=1.000 n=5+5)
EndToEnd/single-col-histogram-range/Prepared-10        322 ± 1%       326 ± 5%    ~     (p=0.563 n=5+5)
EndToEnd/batch-insert-one/Simple-10                  43.7k ± 0%     43.7k ± 0%    ~     (p=0.310 n=5+5)
EndToEnd/batch-insert-one/Prepared-10                43.7k ± 0%     44.3k ± 6%    ~     (p=0.905 n=4+5)
EndToEnd/batch-insert-many/Simple-10                 78.1k ± 6%     78.0k ± 6%    ~     (p=0.548 n=5+5)
EndToEnd/batch-insert-many/Prepared-10               65.2k ± 3%     67.3k ±12%    ~     (p=0.690 n=5+5)
EndToEnd/ored-preds-100/Simple-10                    4.60k ± 0%     4.40k ± 0%  -4.35%  (p=0.029 n=4+4)
EndToEnd/ored-preds-100/Prepared-10                  3.79k ± 0%     3.58k ± 0%  -5.34%  (p=0.016 n=4+5)
EndToEnd/ored-preds-using-params-100/Simple-10       7.84k ± 0%     7.64k ± 0%  -2.56%  (p=0.008 n=5+5)
EndToEnd/ored-preds-using-params-100/Prepared-10     6.90k ± 0%     6.70k ± 0%  -2.89%  (p=0.016 n=4+5)
```